+++ src/main.cpp
@@ -10,6 +10,7 @@ int main(int argc, char *argv[])
   App.setOrganizationName("ChungZH");
   App.setApplicationName("Notepanda");
   App.setApplicationVersion("0.0.1-rc.1");
+
   QApplication::setStyle(QStyleFactory::create("Fusion"));
   MainWindow notepanda;
   notepanda.show();