+++ src/ui/preferenceswindow.cpp
@@ -1,6 +1,9 @@
 #include "preferenceswindow.h"
+#include "ui_preferenceswindow.h"
 
-PreferencesWindow::PreferencesWindow()
+
+PreferencesWindow::PreferencesWindow(QWidget *parent)
+    : QDialog(parent), ui(new Ui::PreferencesWindow)
 {
 
 }