+++ src/ui/preferenceswindow.h
@@ -4,11 +4,18 @@
 #include <QObject>
 #include <QWidget>
 
+#include "ui_preferenceswindow.h"
+
 class PreferencesWindow
+        : public QDialog
+        , private Ui::PreferencesWindow
 {
     Q_OBJECT
 public:
-    PreferencesWindow();
+    PreferencesWindow(QWidget *parent = nullptr);
+
+   private:
+    Ui::PreferencesWindow *ui;
 };
 
 #endif // PREFERENCESWINDOW_H